restructure and rename files
