+++ src/styles/judging-panel.css
